+++ Automate/release-notes.md
@@ -7,7 +7,7 @@ Released 26 November 2019.
 * Updated for Stardew Valley 1.4, including...
   * new machines (fish pond and wood chipper);
   * new recipes (aged roe, caviar, dinosaur mayonnaise, and green tea);
-  * chance of double output with higher quality items for Cheese Press and Loom.
+  * chance of double Loom output with higher quality input.
 * Fixed auto-grabbers not showing empty sprite when emptied automatically.
 
 ## 1.13.2